kernel - Upgrade buffer space tracking variables from int to long
* Several bufspace-related buffer cache parameters can now overflow a
  32 bit integer on machines with large amounts (~64G+) of memory.
  Change these to long.

  bufspace, maxbufspace, maxbufmallocspace, lobufspace, hibufspace,
  lorunningspace, hirunningspace, dirtybufspace,
  dirtybufspacehw, runningbufspace, lodirtybufspace, hidirtybufspace.

* Also requires an API change to libkcore/libkinfo which effects top.

+++ lib/libkcore/kcore.h
@@ -93,7 +93,7 @@ int	 kcore_get_sched_profhz(struct kcore_data *, int *);
 int	 kcore_get_sched_stathz(struct kcore_data *, int *);
 int	 kcore_get_tty_tk_nin(struct kcore_data *, uint64_t *);
 int	 kcore_get_tty_tk_nout(struct kcore_data *, uint64_t *);
-int	 kcore_get_vfs_bufspace(struct kcore_data *, int *);
+int	 kcore_get_vfs_bufspace(struct kcore_data *, long *);
 __END_DECLS;
 
 #endif
