+++ sys/vfs/hammer/hammer.h
@@ -911,8 +911,8 @@ struct hammer_mount {
 	struct hammer_mod_rb_tree data_root;	/* dirty data buffers */
 	struct hammer_mod_rb_tree meta_root;	/* dirty meta bufs    */
 	struct hammer_mod_rb_tree lose_root;	/* loose buffers      */
-	int	locked_dirty_space;		/* meta/volu count    */
-	int	io_running_space;		/* io_token */
+	long	locked_dirty_space;		/* meta/volu count    */
+	long	io_running_space;		/* io_token */
 	int	unused01;
 	int	objid_cache_count;
 	int	dedup_cache_count;
@@ -1038,13 +1038,13 @@ extern int64_t hammer_stats_inode_flushes;
 extern int64_t hammer_stats_commits;
 extern int64_t hammer_stats_undo;
 extern int64_t hammer_stats_redo;
-extern int hammer_count_dirtybufspace;
+extern long hammer_count_dirtybufspace;
 extern int hammer_count_refedbufs;
 extern int hammer_count_reservations;
-extern int hammer_count_io_running_read;
-extern int hammer_count_io_running_write;
+extern long hammer_count_io_running_read;
+extern long hammer_count_io_running_write;
 extern int hammer_count_io_locked;
-extern int hammer_limit_dirtybufspace;
+extern long hammer_limit_dirtybufspace;
 extern int hammer_limit_recs;
 extern int hammer_limit_inode_recs;
 extern int hammer_limit_reclaims;
