+++ usr.bin/top/m_dragonfly.c
@@ -391,7 +391,7 @@ get_system_info(struct system_info *si)
 		static unsigned int swap_delay = 0;
 		static int swapavail = 0;
 		static int swapfree = 0;
-		static int bufspace = 0;
+		static long bufspace = 0;
 
 		if (sysctlbyname("vm.vmstats", &vms, &vms_size, NULL, 0))
 			err(1, "sysctlbyname: vm.vmstats");
